More SectorInterface refactoring.
