+++ lib/vfs/sectorinterface.h
@@ -2,6 +2,7 @@
 #define SECTORINTERFACE_H
 
 class Image;
+class Sector;
 
 class SectorInterface
 {
@@ -10,28 +11,9 @@ public:
 	virtual std::shared_ptr<Sector> put(unsigned track, unsigned side, unsigned sectorId) = 0;
 
 	virtual void flush() {}
-};
-
-class ImageSectorInterface : public SectorInterface
-{
-public:
-	ImageSectorInterface(std::shared_ptr<Image> image):
-		_image(image)
-	{}
 
 public:
-	std::shared_ptr<const Sector> get(unsigned track, unsigned side, unsigned sectorId)
-	{
-		return _image->get(track, side, sectorId);
-	}
-
-	std::shared_ptr<Sector> put(unsigned track, unsigned side, unsigned sectorId)
-	{
-		return _image->put(track, side, sectorId);
-	}
-
-private:
-	std::shared_ptr<Image> _image;
+	static std::unique_ptr<SectorInterface> createImageSectorInterface(std::shared_ptr<Image> image);
 };
 
 #endif
